requirements.txt
joblib==1.3.2
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.4.1.post1
seaborn==0.13.2
streamlit==1.33.0
xgboost==2.0.3
